Force IPv4 ping in ServerStart.bat when using DNS resolution
Ping does not respond with a TTL if the server responds with an IPv6
address. This essentially means that the script will think the user does
not have internet, if they have a public IPv6 address. Forcing IPv4 in
the ping command, ensures TTL gets returned, and the startup can
proceed.

This issue is only relevant in the batch script, as that is the only
script that asserts on TTL being present in the output. It is also only
an issue when pinging minecraftforge.net, as that relies on DNS
resolution. This means the output differs if the DNS server returns IPv4
(the A record) or IPv6 (the AAAA record).

Closes #88